Make scala2-compat changes
