+++ core/src/main/scala/chisel3/BitsImpl.scala
@@ -3,8 +3,9 @@
 package chisel3
 
 import chisel3.experimental.{requireIsHardware, SourceInfo}
-import chisel3.internal.{_resizeToWidth, throwException, BaseModule}
+import chisel3.internal.{throwException, BaseModule}
 import chisel3.internal.Builder.pushOp
+import chisel3.internal.util._
 import chisel3.internal.firrtl.ir._
 import chisel3.internal.firrtl.ir.PrimOp._
 import _root_.firrtl.{ir => firrtlir}
@@ -41,7 +42,7 @@ private[chisel3] trait BitsImpl extends Element { self: Bits =>
   }
 
   protected def _headImpl(n: Int)(implicit sourceInfo: SourceInfo): UInt = {
-    width match {
+    this.width match {
       case KnownWidth(x) => require(x >= n, s"Can't head($n) for width $x < $n")
       case UnknownWidth  => ()
     }
@@ -99,7 +100,7 @@ private[chisel3] trait BitsImpl extends Element { self: Bits =>
   }
 
   protected def _applyImpl(x: UInt)(implicit sourceInfo: SourceInfo): Bool =
-    do_extract(x)
+    extract(x)
 
   protected def _applyImpl(x: Int, y: Int)(implicit sourceInfo: SourceInfo): UInt = {
     if ((x < y && !(x == -1 && y == 0)) || y < 0) {
@@ -193,9 +194,9 @@ private[chisel3] trait BitsImpl extends Element { self: Bits =>
   protected def _asSIntImpl(implicit sourceInfo: SourceInfo): SInt
 
   protected def _asBoolImpl(implicit sourceInfo: SourceInfo): Bool = {
-    width match {
+    this.width match {
       case KnownWidth(1) => this(0)
-      case _             => throwException(s"can't covert ${this.getClass.getSimpleName}$width to Bool")
+      case _             => throwException(s"can't covert ${this.getClass.getSimpleName}${this.width} to Bool")
     }
   }
 
@@ -330,15 +331,15 @@ private[chisel3] trait UIntImpl extends BitsImpl with Num[UInt] { self: UInt =>
   protected def _rotateLeftImpl(n: Int)(implicit sourceInfo: SourceInfo): UInt = width match {
     case _ if (n == 0)             => this
     case KnownWidth(w) if (w <= 1) => this
-    case KnownWidth(w) if n >= w   => do_rotateLeft(n % w)
-    case _ if (n < 0)              => do_rotateRight(-n)
+    case KnownWidth(w) if n >= w   => rotateLeft(n % w)
+    case _ if (n < 0)              => rotateRight(-n)
     case _                         => tail(n) ## head(n)
   }
 
   protected def _rotateRightImpl(n: Int)(implicit sourceInfo: SourceInfo): UInt = width match {
-    case _ if (n <= 0)             => do_rotateLeft(-n)
+    case _ if (n <= 0)             => rotateLeft(-n)
     case KnownWidth(w) if (w <= 1) => this
-    case KnownWidth(w) if n >= w   => do_rotateRight(n % w)
+    case KnownWidth(w) if n >= w   => rotateRight(n % w)
     case _                         => this(n - 1, 0) ## (this >> n)
   }
